+++ src/core/search/comparators/int_comparator.rs
@@ -70,8 +70,8 @@ impl IntComparator {
 
 impl FieldComparator for IntComparator {
     type V = i32;
-    fn compare(&self, slot1: i32, slot2: i32) -> i32 {
-        self.values[slot1 as usize].cmp(&self.values[slot2 as usize]) as i32
+    fn compare(&self, slot1: usize, slot2: usize) -> i32 {
+        self.values[slot1].cmp(&self.values[slot2]) as i32
     }
 
     fn set_top_value(&mut self, value: Self::V) {
@@ -79,8 +79,8 @@ impl FieldComparator for IntComparator {
         self.top_value = value;
     }
 
-    fn value(&self, slot: i32) -> Option<Self::V> {
-        Some(self.values[slot as usize])
+    fn value(&self, slot: usize) -> Option<Self::V> {
+        Some(self.values[slot])
     }
 
     type LeafFieldComparator<LR>
