+++ src/core/search/sorted_numeric_sort_field.rs
@@ -562,7 +562,7 @@ macro_rules! impl_sorted_numeric_comparator {
         impl FieldComparator for $name {
             type V = <$base as FieldComparator>::V;
 
-            fn compare(&self, slot1: i32, slot2: i32) -> i32 {
+            fn compare(&self, slot1: usize, slot2: usize) -> i32 {
                 self.base.compare(slot1, slot2)
             }
 
@@ -570,7 +570,7 @@ macro_rules! impl_sorted_numeric_comparator {
                 self.base.set_top_value(value);
             }
 
-            fn value(&self, slot: i32) -> Option<Self::V> {
+            fn value(&self, slot: usize) -> Option<Self::V> {
                 self.base.value(slot)
             }
